Handle simple expressions
Negative numbers are parsed as unary minus expression, and some of the
fields are calculated as well, e.g. `time = 5 * 60`

Reduces the number of `null`s in the data (#1) and at least partially
fixes #2.

diffstat for 0.15.23 data files before and after below (js files omitted
for brevity.)

 decorative.json |  186 !!!!!!
 entity.json     | 1590 ++++!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 equipment.json  |    3
 item.json       |   33 !
 legacy.json     |    4
 other.json      |    9
 technology.json |    2
 14 files changed, 306 insertions(+), 3348 modifications(!)

The following AST types are still unhandled and result in null:
- CallExpression
- Identifier
- LogicalExpression
- MemberExpression
- TableCallExpression